Modify the applications tab to use breadcrumbs
FlipScrollView can now be controlled from the outside to set which
item is used as its root view. Also now it notifies the outside of
root changes.

At the other end, the launcher now adds an extra row of buttons at the
top of the application tab to create a breadcrumb bar. The buttons of
this bar simply ask the FlipScrollView to jump at the right place in
the model.