Fix activity creation based on templates
Found out the activity creation based on templates was broken in two
ways since the latest adjustments:
 * creating pre-created activities wasn't executing the layout.js
   script;
 * creating activities from the "new activity" templates wasn't
   executing the startup apps.

So I just factored as much as possible of the activity creation
behavior into PlasmaApp to avoid such issues. Also ActivityIcon was
using a wrong path for the layout.js script, so passing the right one
now.