Fix panel controller width when running on multiple monitors.
The panel controller was taking only a portion of the bigger screen
width (that was primary). With this patch, we force the widget to
calculate its size before being shown.

Also, syncToLocation needed a small tweak, since when updating
struts were needed (this is, dragging the panel from horizontal to
vertical [or reverse] layout), the panel itself was taking the whole
screen (because of setting the maximum size to QWIDGETSIZE_MAX).

REVIEW: 101254