Added lightDM, it is just like GDM (ConsoleKit) + custom startReserve
Added support for lightDM, which basically is managed like GDM but with
a custom startReserve method (using lightDM dbus interface).
LightDM is under heavy development right now, though it is fully working
it has some API's that may change in the near future (for example now it
identify itself as "GDMSESSION", something that is expected to be fixed
soon.

As final note, there is Qt greeter for lightDM, and David Edmundson is
already working on a KDE one.

CCMAIL: ossi@kde.org