Create XPixmaps for shadow on graphicssystem raster
With graphicssystem raster a QPixmap is internally a QImage causing
the handle to be 0. In order to pass the pixmaps ids to kwin we need
to create an XPixmap first and create the QPixmap from it to force
to native rendering.
It's pretty ugly but there seems to be no other way :-(
REVIEW: 101367