Add UnifyOutputs option into the KRandR KCM too.
KRandRTray and the KCM share the same lib to manage screens, and that
lib is coupled with "Screen management" and configuraiton. There is one
KRandRTray configuration option that allows you to unify the outputs,
and if actived it will force the Output chagnes to be equal. If this
option is enabled, no matter what you configure in the KCM (position,
rotation, resolution...) that it won't work since the library behind
the scenes will make sure to "unify" the options.

These configuration coupling is so deep down the library that I don't
want to mess with it for a minor release, so at the moment what I've done
and what this commit does is add the "unify outputs" options into the KCM
and modify it so the user WON'T be able to setup a configuration
incompatible with "Unify Output".
(cherry picked from commit 1d56786e49937d4d6b599d6a055e8408aa16f1eb)