Use the first (highest) resolution in common instead of last used
The last used resolution can be wrong, can be incompatible or can be
just "ugly" if you hardware setup has changed. This is the common
use case for KRandRTray when you use it to setup quickly and easy a new
monitor.
To fix this wrong behavior (which may end in an unusable system) instead
of loading the last used resolution we're not loading the highest
resolution in common.

As a practical case, my laptop has 1280x800 and my LCD is 1920x1080, for
an unknow reason my las used resolution was 1024x768 which is not
working at all in either of the screens.

Finally I just want to note that this is the behavior that any RandR
user expect, unify outputs at the highest (better) resolution.
(cherry picked from commit 717b4f18621da0355319e03c79dd5dc0576b2b51)