Make RandROutput handle events again so the information is updated
KRandRTray is NOT updating the menu correctly because the event handling
is disabled in RandROutput, this patch fixes that.

Seems that RandROutput eventHandler caused some problems in the past that
now I can't reproduce on 3 systems playing with it for 2 days. Also this
patch makes RandROutput loadConfig if the output is connected/disconnected