Check dialog result
The intention of the previous use of signal "finished()" was to also
delete the dialog on Cancel. But when the dialog closes because the
widget is deleted, we should not try to load/save the script. The fix
is to connect to "finished(int result)", now both for load and save.
CCBUG: 278120