Do NOT apply any new CRTC in handleEven, avoids unexpected results
When we're handling external events we DO NOT want KRandR to interfere,
if an external application did something all we have to do is refresh
our data.

This fixes that when you do:
xrandr --output LVDS1 --off KRandR will decide to turn off all
connected screens if unifyMode is active.