kstartupconfig: Fix return value conversion of system()
Kurt Hindenburg reported that this file did not compile on Mac OS X due
to a compiler error.

Turns out to be some kind of macro magic implementing WEXITSTATUS() on
the return value of system(3). But the return value of system(3) is
already the WEXITSTATUS so we don't need to use that macro here.