make TaskItem return a Task* instead of a smart pointer (KShardPtr).
TaskItem holds onto a QWeakPointer internally, and then is constantly handing out
new KSharedPtr's on that pointer's data. it's used infrequently and where it is used
items should be providing their own smart pointer wrappers on an as-need basis.

TaskItem is based on the idea that the task may go away at any time ... so respect
that contract and do not create new strong shared pointers from its internal pointer.

i believe this to be related to BR272495

CCBUG:272495