use a QWeakPointer to track the lifespan of the task
the itemRemoved signal from TaskGroup has dangling pointers, apparently
by design (the next thing on my list to investigate), and the signal is
delayed. so between the actual deletion of the task and the signal, there
pointer is a dangler. easy solution is to just track it in the WindowTaskItem
itself with a cheap QWeakPointer.

BUG:272495

Conflicts:

	plasma/desktop/applets/tasks/windowtaskitem.cpp