Make Oxygen mouse cursors (pseudo) scalable
This patch makes the mouse cursor themes that KDE provides pseudo
scalable. This means that it removes the "big" themes. Instead of this,
it provides only 1 theme per color. This theme is a multi-size theme.
It contains coursers in 24 (old "normal" size), 48 (old "big" size)
and 72 (new size).

At the same time, it substitutes the duplicate curser images (there
were many of them) by symbolic links to save memory. This way, the
file size of the themes is (even with adding the "72" size) decreased.

The patch modifies kde-workspace/cursors/src which provides the sources
that produce the binary image files.

The patch also modifies the (binary) themes in kde-workspace/cursors
directly.

I hope that GIT manages well all the binary stuff and the symbolic
links ...
BUG: 273936
FIXED-IN: 4.8
REVIEW: 102524