Prefer KIconLoader::Small icon sizes for window thumbnails when possible
This makes the pager actually abide by the system settings designated
Small icon size. So increasing it to e.g. 42x42 would make the window
icons in the pager use that size.

It's set to use the width/height (whichever is smallest), and the
KIconLoader::Small. Whichever value is smaller.

Was previously hardcoded to 16x16. Should be a safe change, going to
backport.
BUG:25766