kwin: fixing high cpu usage in the fade effect
This is a condensed version of Martin's patch that fixes a high cpu
usage in KWin and X. It seems to be due to a window being created and
deleted at almost the same time, such that the fade effect never quits.
The bug was reliably reproducable with starting Amarok.

CCBUG: 290025
CCBUG: 288948