Only checkInhibition if there is a Lid present
This checkInhibition code is designed to inhibit events ONLY when used
in a laptop and it allows to inhibit suspension and other events that
are triggered when the lid  is closed.

So, to avoid affecting desktop pc's and other form factors we are
checking if a lid is present.

Something more global will have to be add in order to support other form
factors such Active devices.