rename setscreensaverinhibited.
to inhibitScreensaver/uninhibitScreensaver. It's better imho. Also had a
note, to keep in mind that it is in fact 'uninhibitScreensaver'. i.e.
it is NOT 'unInhibitScreensaver' counter to the f.d.o spec. honestly I
thought it was a stupid distinction, and here it'll just lead to service
method errors in calling...guarantee it.