add powermanagement engine un/inhibit powermanagement functionality
note that it doesn't actually do anything useful. apparently this is a
bug, of which a GCI student is hacking on.

in theory it should work out of the box when it's fixed...if my d-bus
calls are great. (e.g. qdbusviewer calling pm inhibit does nothing
even).

so now my applet should be good to go, well..after this is resolved.