properly map rblk and wblk in and out of canonical form
the config dialog now loads and saves properly. also, i don't think
the write I/O was being read properly or at all. so now that's fixed as
well.

looks like it works on a desktop quite well, so it's nearly done, just
have to fix and test it when it's in the panel