If kernel newer than 2.6.36 use the backlight type instead of whitelist
If the kernel is newer than 2.6.36 it indicates the type of the
backlight wether it is a firmware/platform or raw. We should
attempt to connect to these interfaces in this order.

This should fix any compatibility issue we have right now and make this
code comaptible with drivers that may appear in the future.

BUG: 288180
CCBUG: 292873
CCMAIL: lukas@kde.org
FIXED-IN: 4.9