Do not use GraphicsScene to calculate Thumbnail position
Instead of using the scene it is better to use the toplevel
item's parent position. This works better in case the scene
is endless and the toplevel item is not positioned at 0/0 as
it is the case for anything from Plasma.

There are still slight layouting issues which I think is
caused by thumbnails not honouring the margin. Will investigate.

REVIEW: 104395